+++ lib/geometries/arkit_sphere.dart
@@ -1,13 +1,20 @@
+import 'package:arkit_plugin/geometries/arkit_geometry.dart';
+import 'package:arkit_plugin/geometries/arkit_material.dart';
 import 'package:arkit_plugin/geometries/arkit_position.dart';
 import 'package:meta/meta.dart';
 
-class ARKitSphere {
-  ARKitSphere({@required this.position, @required this.radius});
+class ARKitSphere extends ARKitGeometry {
+  ARKitSphere({
+    @required ARKitPosition position,
+    @required this.radius,
+    List<ARKitMaterial> materials,
+  }) : super(position, materials: materials);
 
-  final ARKitPosition position;
   final double radius;
 
-  Map<String, dynamic> toMap() {
-    return <String, dynamic>{'radius': radius, 'position': position.toMap()};
-  }
+  Map<String, dynamic> toMap() => <String, dynamic>{
+        'radius': radius,
+        'position': position.toMap(),
+        'materials': materials.map((m) => m.toMap()).toList(),
+      };
 }